Move view controller and app switch delegates
- Definitions for BTViewControllerPresentingDelegate and
  BTAppSwitchDelegate moved to more appropriate header files.
