--- BraintreeCore/BTDelegates.m
@@ -1,7 +1,0 @@
-#import <Foundation/Foundation.h>
-
-NSString * const BTAppSwitchWillSwitchNotification = @"com.braintreepayments.BTAppSwitchWillSwitchNotification";
-NSString * const BTAppSwitchDidSwitchNotification = @"com.braintreepayments.BTAppSwitchDidSwitchNotification";
-NSString * const BTAppSwitchWillProcessPaymentInfoNotification = @"com.braintreepayments.BTAppSwitchWillProcessPaymentInfoNotification";
-
-NSString * const BTAppSwitchNotificationTargetKey = @"BTAppSwitchNotificationTargetKey";
