+++ BraintreeCore/Public/BTAppSwitch.h
@@ -1,33 +1,10 @@
 #import <Foundation/Foundation.h>
 
+@protocol BTAppSwitchHandler;
+
 NS_ASSUME_NONNULL_BEGIN
 
-/// @protocol BTAppSwitchHandler
-/// @description A protocol for handling the return from switching out of an app to gather payment information.
-/// The app may switch out to Mobile Safari or to a native app.
-@protocol BTAppSwitchHandler
-
-@required
-
-/// Determine whether the app switch return URL can be handled.
-///
-/// @param url the URL you receive in `application:openURL:sourceApplication:annotation` when returning to
-/// your app
-/// @param sourceApplication The source application you receive in `application:openURL:sourceApplication:annotation`
-/// @return `YES` when the object can handle returning from the application with a URL
-+ (BOOL)canHandleAppSwitchReturnURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication;
-
-/// Pass control back to `BTPayPalDriver` after returning from app or browser switch.
-///
-/// @param url The URL you receive in `application:openURL:sourceApplication:annotation`
-+ (void)handleAppSwitchReturnURL:(NSURL *)url;
-
-@optional
-
-/// Indicates whether an iOS app is installed and available for app switch.
-- (BOOL)isiOSAppAvailableForAppSwitch;
-
-@end
+#pragma mark - BTAppSwitch
 
 /// @class BTAppSwitch
 /// @description Handles return URLs when returning from app switch and routes the return URL to the correct
@@ -70,6 +47,92 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// Unregisters a class that knows how to handle a return from app switch
 - (void)unregisterAppSwitchHandler:(Class<BTAppSwitchHandler>)handler;
+
+@end
+
+#pragma mark - BTAppSwitchDelegate
+
+extern NSString * const BTAppSwitchWillSwitchNotification;
+extern NSString * const BTAppSwitchDidSwitchNotification;
+extern NSString * const BTAppSwitchWillProcessPaymentInfoNotification;
+extern NSString * const BTAppSwitchNotificationTargetKey;
+
+/// Specifies the destination of an app switch
+typedef NS_ENUM(NSInteger, BTAppSwitchTarget) {
+    BTAppSwitchTargetUnknown = 0,
+    /// Native app
+    BTAppSwitchTargetNativeApp,
+    /// Browser (i.e. Mobile Safari)
+    BTAppSwitchTargetWebBrowser,
+};
+
+/// Protocol for receiving payment lifecycle messages from a payment option
+/// that may initiate an app or browser switch event to authorize payments.
+@protocol BTAppSwitchDelegate <NSObject>
+
+/// The app switcher will perform an app switch in order to obtain user
+/// payment authorization.
+///
+/// Your implementation of this method may set your app to the state
+/// it should be in if the user manually app-switches back to your app.
+/// For example, re-enable any controls that are disabled.
+///
+/// @param appSwitcher The app switcher
+- (void)appSwitcherWillPerformAppSwitch:(id)appSwitcher;
+
+/// Delegates receive this message when the app switcher has successfully performed an app switch.
+///
+/// You may use this hook to prepare your UI for app switch return. Keep in mind that
+/// users may manually switch back to your app via the iOS task manager.
+///
+/// @note You may also hook into the app switch lifecycle via UIApplicationWillResignActiveNotification.
+///
+/// @param appSwitcher The app switcher instance performing user authentication
+/// @param target The destination that was actually used for this app switch
+- (void)appSwitcher:(id)appSwitcher didPerformSwitchToTarget:(BTAppSwitchTarget)target;
+
+/// The app switcher has obtained user payment details and/or user
+/// authorization and will process the results.
+///
+/// This typically indicates asynchronous network activity.
+/// When you receive this message, your UI should indicate activity.
+///
+/// In the case of an app switch, this message indicates that the user has returned to this app;
+/// this is usually after handleAppSwitchReturnURL: is called in your UIApplicationDelegate.
+///
+/// @note You may also hook into the app switch lifecycle via UIApplicationWillResignActiveNotification.
+///
+/// @param appSwitcher The app switcher
+- (void)appSwitcherWillProcessPaymentInfo:(id)appSwitcher;
+
+@end
+
+#pragma mark - BTAppSwitchHandler protocol
+
+/// @protocol BTAppSwitchHandler
+/// @description A protocol for handling the return from switching out of an app to gather payment information.
+/// The app may switch out to Mobile Safari or to a native app.
+@protocol BTAppSwitchHandler
+
+@required
+
+/// Determine whether the app switch return URL can be handled.
+///
+/// @param url the URL you receive in `application:openURL:sourceApplication:annotation` when returning to
+/// your app
+/// @param sourceApplication The source application you receive in `application:openURL:sourceApplication:annotation`
+/// @return `YES` when the object can handle returning from the application with a URL
++ (BOOL)canHandleAppSwitchReturnURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication;
+
+/// Pass control back to `BTPayPalDriver` after returning from app or browser switch.
+///
+/// @param url The URL you receive in `application:openURL:sourceApplication:annotation`
++ (void)handleAppSwitchReturnURL:(NSURL *)url;
+
+@optional
+
+/// Indicates whether an iOS app is installed and available for app switch.
+- (BOOL)isiOSAppAvailableForAppSwitch;
 
 @end
